+++ frontend/app/community/page.tsx
@@ -1,35 +1,19 @@
-'use client';
+import { Metadata } from "next";
+import CommunityPage from "./CommunityClient";
 
-import React from 'react';
-import Navbar from '../components/Navbar';
-import Footer from '../components/Footer';
-import CommunityHero from './components/CommunityHero';
-import CommunityStats from './components/CommunityStats';
-import ProposalsList from './components/ProposalsList';
-import SocialFeeds from './components/SocialFeeds';
-import CommunityGuidelines from './components/CommunityGuidelines';
+export const metadata: Metadata = {
+  title: "Community & Governance",
+  description: "Join the Nestera community. Participate in governance, propose new features, and help shape the future of decentralized savings.",
+  alternates: {
+    canonical: "/community",
+  },
+  openGraph: {
+    title: "Community & Governance - Nestera",
+    description: "Join the Nestera community. Participate in governance, propose new features, and help shape the future of decentralized savings.",
+    images: ["/api/og?page=community"],
+  },
+};
 
-export default function CommunityPage() {
-  return (
-    <main className="min-h-screen bg-[#061a1a]">
-      <Navbar />
-      
-      <CommunityHero />
-      
-      <div className="relative">
-        {/* Horizontal divider with gradient */}
-        <div className="w-full h-px bg-gradient-to-r from-transparent via-white/10 to-transparent" />
-        <CommunityStats />
-        <div className="w-full h-px bg-gradient-to-r from-transparent via-white/10 to-transparent" />
-      </div>
-
-      <ProposalsList />
-      
-      <SocialFeeds />
-      
-      <CommunityGuidelines />
-
-      <Footer />
-    </main>
-  );
+export default function Page() {
+  return <CommunityPage />;
 }
