+++ frontend/app/components/JsonLd.tsx
@@ -1,0 +1,16 @@
+import React from 'react';
+
+interface JsonLdProps {
+  data: Record<string, any>;
+}
+
+const JsonLd: React.FC<JsonLdProps> = ({ data }) => {
+  return (
+    <script
+      type="application/ld+json"
+      dangerouslySetInnerHTML={{ __html: JSON.stringify(data) }}
+    />
+  );
+};
+
+export default JsonLd;
