+++ frontend/app/docs/page.tsx
@@ -1,21 +1,19 @@
-'use client';
+import { Metadata } from "next";
+import DocsPage from "./DocsClient";
 
-import React, { useState } from 'react';
-import DocsSidebar, { DocSection } from './components/DocsSidebar';
-import DocsSections from './components/DocsSections';
+export const metadata: Metadata = {
+  title: "Documentation",
+  description: "Learn how to use Nestera, understand the technology behind our decentralized savings platform, and find answers to your questions.",
+  alternates: {
+    canonical: "/docs",
+  },
+  openGraph: {
+    title: "Documentation - Nestera",
+    description: "Learn how to use Nestera, understand the technology behind our decentralized savings platform, and find answers to your questions.",
+    images: ["/api/og?page=docs"],
+  },
+};
 
-export default function DocsPage() {
-  const [activeSection, setActiveSection] = useState<DocSection>('getting-started');
-
-  return (
-    <div className="flex flex-col md:flex-row gap-12 min-h-[600px]">
-      <DocsSidebar 
-        activeSection={activeSection} 
-        onSectionChange={setActiveSection} 
-      />
-      <div className="flex-1 max-w-4xl">
-        <DocsSections section={activeSection} />
-      </div>
-    </div>
-  );
+export default function Page() {
+  return <DocsPage />;
 }
